requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
WTForms==3.1.1
email-validator>=2.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow>=10.2.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
gunicorn==21.2.0
#psycopg2-binary==2.9.9  # Optional: Only needed for PostgreSQL
flask-cors==4.0.0
cryptography>=41.0.0

